refactor: split commands/mod.rs into config_cmds, context_cmds, memory, session
commands/mod.rs had grown past 3,000 lines. Move the command handlers into
four focused submodules — config_cmds (/config, /backend, /model, /verbose…),
context_cmds (/context, /compact, /reset, /checkpoints), memory (/index,
/map, /memory, /remember, /forget), and session (/new, /undo, /session,
/resume, /export, /path) — leaving dispatch and the command list in mod.rs.
No behavior change.
